Fix SelectSelector to be compatible with monkey_patch
When using eventlet.monkey_patch(), it disables all the Selectors but
the SelectSelector. Another thing it does is modify how the constructor
handles the class methods.
The way the SelectSelector handles the _select function is incompatible
with that code.
We should have a bound method calling the select function (as is done
for the Win32) in order to fix this compatibility issue.